2018 08 31 14.09 增加了用户列表查看页，对页面进行了一下样式优化，改掉了一些bug，新加照片审核功能（在员工审核时进行，照片审核不通过，人员信息会录入到云端，但照片会清空需要重新上传）

+++ src/main/java/com/thinkgem/jeesite/modules/sys/service/SystemService.java
@@ -158,7 +158,7 @@ public class SystemService extends BaseService implements InitializingBean {
 			user.preUpdate();
 			userDao.update(user);
 		}
-		if (user.getLoginFlag().equals("0")&&!user.getIssh().equals("2")&&StringUtils.isNotBlank(user.getId())){
+		if (user.getLoginFlag().equals("0")&&StringUtils.isNotEmpty(user.getIssh())&&!user.getIssh().equals("2")&&StringUtils.isNotBlank(user.getId())){
 			// 更新用户与角色关联
 			userDao.deleteUserRole(user);
 			if (user.getRoleList() != null && user.getRoleList().size() > 0){
